% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHubs.R
\name{detectHubs}
\alias{detectHubs}
\title{Detect the interaction hub}
\usage{
detectHubs(interaction, pval_cutoff = 0.05, ...)
}
\arguments{
\item{interaction}{An object of \link[S4Vectors:Pairs-class]{Pairs} to
represent interactions.}

\item{pval_cutoff}{Cutoff P value for interaction hub by poisson distribution}

\item{...}{Not used.}
}
\value{
A list of interaction hubs with elements:
 hub_connection, Pairs object represent interactions interacted with hubs;
 hubs, GRanges object represent regions involved in hubs;
 hub_regions, GRanges object represent regions interacted with hubs.
}
\description{
Define the interaction hub from input Pairs.
}
\examples{
library(rtracklayer)
p <- system.file("extdata", "WT.2.bedpe",
                 package = "GenomicInteractionHub")
interactions <- import(con=p, format="bedpe")
hubs <- detectHubs(interactions)
}
